requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
databases==0.9.0
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
limits==4.2
loguru==0.7.3
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
murmurhash==1.0.12
nltk==3.9.1
nodeenv==1.9.1
numpy==2.0.2
openai==1.74.0
packaging==24.2
pamqp==3.3.0
pendulum==3.0.0
pika==1.3.2
pluggy==1.5.0
preshed==3.0.9
prisma==0.15.0
propcache==0.3.1
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9
secure==0.3.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.46.1
StrEnum==0.4.15
textblob==0.19.0
thinc==8.3.6
time-machine==2.16.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.0
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
yarl==1.19.0
